add setup scripts
