+++ .conf
@@ -1,0 +1,8 @@
+weightsdir=modelweights
+rawweights=modelweights/yolov3.weights
+raww=yolov3.weights
+weightsurl=https://pjreddie.com/media/files/yolov3.weights
+modelweights=modelweights/model.h5
+buildscript=model/create_model.py
+buildlog=model-build.log
+testimg=test/zebra.jpg
