+++ README.md
@@ -1,0 +1,17 @@
+# YOLOv3
+
+## USAGE
+
+To download weights and setup model:
+  - Windows users: run `setup.bat`
+  - MacOS or UNIX/LINUX users: run `setup.sh`
+  - Windows users with git bash can also run `setup.sh`
+
+This step will download the weights, build the model and test setup. 
+
+To feed images from web camera:
+  - Windows users: run `main.bat`
+  - MacOS or UNIX/LINUX users: run `main.sh`
+  - Windows users with git bash can also run `main.sh`
+
+Use `--no-nms` to disable non-max suppression and increase FPS.
