+++ main.bat
@@ -1,0 +1,1 @@
+python bin\from_camera.py %*
