+++ model/yolov3.py
@@ -1,5 +1,3 @@
-import struct
-import numpy as np
 from keras.layers import Conv2D
 from keras.layers import Input
 from keras.layers import BatchNormalization
@@ -8,6 +6,25 @@ from keras.layers import ZeroPadding2D
 from keras.layers import UpSampling2D
 from keras.layers.merge import add, concatenate
 from keras.models import Model
+
+LABELS = [
+    "person", "bicycle", "car", "motorbike", "aeroplane", "bus", "train", "truck",
+    "boat", "traffic light", "fire hydrant", "stop sign", "parking meter", "bench",
+    "bird", "cat", "dog", "horse", "sheep", "cow", "elephant", "bear", "zebra", "giraffe",
+    "backpack", "umbrella", "handbag", "tie", "suitcase", "frisbee", "skis", "snowboard",
+    "sports ball", "kite", "baseball bat", "baseball glove", "skateboard", "surfboard",
+    "tennis racket", "bottle", "wine glass", "cup", "fork", "knife", "spoon", "bowl", "banana",
+    "apple", "sandwich", "orange", "broccoli", "carrot", "hot dog", "pizza", "donut", "cake",
+    "chair", "sofa", "pottedplant", "bed", "diningtable", "toilet", "tvmonitor", "laptop", "mouse",
+    "remote", "keyboard", "cell phone", "microwave", "oven", "toaster", "sink", "refrigerator",
+    "book", "clock", "vase", "scissors", "teddy bear", "hair drier", "toothbrush"
+    ]
+
+ANCHORS = [
+    [116, 90, 156, 198, 373, 326],
+    [30, 61, 62, 45, 59, 119],
+    [10, 13, 16, 30, 33, 23]
+    ]
 
 
 def _conv_block(inp, convs, skip=True):
@@ -130,58 +147,3 @@ def make_yolov3_model():
                                {'filter': 255, 'kernel': 1, 'stride': 1, 'bnorm': False, 'leaky': False, 'layer_idx': 105}], skip=False)
     model = Model(input_image, [yolo_82, yolo_94, yolo_106])
     return model
-
-
-class WeightReader:
-    def __init__(self, weight_file):
-        with open(weight_file, 'rb') as w_f:
-            major,    = struct.unpack('i', w_f.read(4))
-            minor,    = struct.unpack('i', w_f.read(4))
-            revision, = struct.unpack('i', w_f.read(4))
-            if (major * 10 + minor) >= 2 and major < 1000 and minor < 1000:
-                w_f.read(8)
-            else:
-                w_f.read(4)
-            transpose = (major > 1000) or (minor > 1000)
-            binary = w_f.read()
-        self.offset = 0
-        self.all_weights = np.frombuffer(binary, dtype='float32')
-
-    def read_bytes(self, size):
-        self.offset = self.offset + size
-        return self.all_weights[self.offset - size:self.offset]
-
-    def load_weights(self, model):
-        for i in range(106):
-            try:
-                conv_layer = model.get_layer('conv_' + str(i))
-                print("loading weights of convolution #" + str(i))
-                if i not in [81, 93, 105]:
-                    norm_layer = model.get_layer('bnorm_' + str(i))
-                    size = np.prod(norm_layer.get_weights()[0].shape)
-                    beta = self.read_bytes(size)  # bias
-                    gamma = self.read_bytes(size)  # scale
-                    mean = self.read_bytes(size)  # mean
-                    var = self.read_bytes(size)  # variance
-                    weights = norm_layer.set_weights([gamma, beta, mean, var])
-                if len(conv_layer.get_weights()) > 1:
-                    bias = self.read_bytes(
-                        np.prod(conv_layer.get_weights()[1].shape))
-                    kernel = self.read_bytes(
-                        np.prod(conv_layer.get_weights()[0].shape))
-                    kernel = kernel.reshape(
-                        list(reversed(conv_layer.get_weights()[0].shape)))
-                    kernel = kernel.transpose([2, 3, 1, 0])
-                    conv_layer.set_weights([kernel, bias])
-                else:
-                    kernel = self.read_bytes(
-                        np.prod(conv_layer.get_weights()[0].shape))
-                    kernel = kernel.reshape(
-                        list(reversed(conv_layer.get_weights()[0].shape)))
-                    kernel = kernel.transpose([2, 3, 1, 0])
-                    conv_layer.set_weights([kernel])
-            except ValueError:
-                print("no convolution #" + str(i))
-
-    def reset(self):
-        self.offset = 0
