+++ requirements.txt
@@ -1,0 +1,5 @@
+tensorflow==1.14.0
+numpy==1.17.2
+opencv-python==4.1.1.26
+Keras==2.3.1
+matplotlib==3.1.1
